refactor: replace SvelteKit with vanilla JS
Drop the entire SvelteKit toolchain in favor of plain ES modules
and Vite as a minimal bundler. The app logic, boot sequence, section
navigation, and game launcher are now a single vanilla JS module
using JSDoc types, event delegation, and modern ECMA patterns.
No runtime framework dependencies remain.
