+++ .gitignore
@@ -1,11 +1,6 @@
 node_modules
 
 # Output
-.output
-.vercel
-.netlify
-.wrangler
-/.svelte-kit
 /build
 
 # OS
