+++ jsconfig.json
@@ -1,5 +1,4 @@
 {
-	"extends": "./.svelte-kit/tsconfig.json",
 	"compilerOptions": {
 		"allowJs": true,
 		"checkJs": true,
@@ -9,6 +8,9 @@
 		"skipLibCheck": true,
 		"sourceMap": true,
 		"strict": true,
-		"moduleResolution": "bundler"
-	}
+		"module": "es2022",
+		"moduleResolution": "bundler",
+		"target": "es2022"
+	},
+	"include": ["src/**/*.js"]
 }
