+++ package.json
@@ -4,25 +4,16 @@
 	"version": "4.1.1",
 	"type": "module",
 	"scripts": {
-		"dev": "vite dev",
+		"dev": "vite",
 		"build": "vite build",
 		"preview": "vite preview",
-		"check": "svelte-kit sync && svelte-check --tsconfig ./jsconfig.json",
-		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./jsconfig.json --watch",
 		"deploy": "npm run build && touch build/.nojekyll && gh-pages -d build",
 		"version:patch": "npm version patch && git push && git push --tags",
 		"version:minor": "npm version minor && git push && git push --tags",
 		"version:major": "npm version major && git push && git push --tags"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/adapter-static": "^3.0.6",
-		"@sveltejs/kit": "^2.9.0",
-		"@sveltejs/vite-plugin-svelte": "^5.0.0",
 		"gh-pages": "^6.2.0",
-		"svelte": "^5.0.0",
-		"svelte-check": "^4.0.0",
-		"typescript": "^5.0.0",
 		"vite": "^6.0.0"
 	}
 }
