--- svelte.config.js
@@ -1,22 +1,0 @@
-import adapter from '@sveltejs/adapter-static';
-
-/** @type {import('@sveltejs/kit').Config} */
-const config = {
-	kit: {
-		alias: {
-			$lib: 'src/lib',
-		},
-		adapter: adapter({
-			pages: 'build',
-			assets: 'build',
-			fallback: 'index.html',
-			precompress: false,
-			strict: true
-		}),
-		paths: {
-			base: ''
-		}
-	}
-};
-
-export default config;
