+++ vite.config.js
@@ -1,11 +1,8 @@
-import { sveltekit } from '@sveltejs/kit/vite';
 import { defineConfig } from 'vite';
 
 export default defineConfig({
-	plugins: [sveltekit()],
-	server: {
-		fs: {
-			allow: ['..']
-		}
-	}
+	publicDir: 'static',
+	build: {
+		outDir: 'build',
+	},
 });
